requirements.txt
python == 3
tkinter >= 8.6.6
pillow >=5.2.0